refactor: rename QueryDishMediaResponse to SearchDishMediaResponse and update related mappings
