+++ app-expo/app/[locale]/(tabs)/_layout.tsx
@@ -64,6 +64,7 @@ export default function TabLayout() {
 					tabBarIcon: ({ size, color }) => <User size={ICON_SIZE} color={color} />,
 				}}
 			/>
+			<Tabs.Screen name="posts" options={{ href: null }} />
 		</Tabs>
 	);
 }
