+++ shared/api/v1/res/dish-media.response.ts
@@ -22,12 +22,12 @@ export type DishMediaEntry = {
 };
 
 /** GET /v1/dish-media/search のレスポンス型 */
-export type QueryDishMediaResponse = DishMediaEntry[];
+export type SearchDishMediaResponse = DishMediaEntry[];
 
 /** GET /v1/dish-media?ids=... のレスポンス型 */
 export type QueryDishMediaByIdsResponse = {
-        items: DishMediaEntry[];
-        notFound: string[];
+	items: DishMediaEntry[];
+	notFound: string[];
 };
 
 /** POST /v1/dish-media のレスポンス型 */
